separate the authentication endpoints and the files endpoints; And add documentation basic
